+++ packages/client/src/i18n/messages/form.ts
@@ -222,6 +222,16 @@ export const formMessageDescriptors = {
     description: 'Label for father select option to certify record',
     id: 'form.field.label.app.certifyRecordTo.father'
   },
+  certifyRecordToGroom: {
+    defaultMessage: 'Print and issue to groom',
+    description: 'Label for groom select option to certify record',
+    id: 'form.field.label.app.certifyRecordTo.groom'
+  },
+  certifyRecordToBride: {
+    defaultMessage: 'Print and issue to bride',
+    description: 'Label for bride select option to certify record',
+    id: 'form.field.label.app.certifyRecordTo.bride'
+  },
   typeOfId: {
     defaultMessage: 'Type of ID',
     description: "Input label for certificate collector's id type options",